refactor(engine): split the local-copy context state module by concern
context_impl/state.rs had grown to 1549 lines holding four unrelated
concerns, and CopyContext::new was a 212-line constructor in which 130
lines derived batch and compression sub-components inline.

Split along concern boundaries, following the decomposition convention
already used for context_impl/options/:

  batch_setup.rs       - one builder per batch/compression sub-component,
                         so new() assembles rather than derives
  link_dest.rs         - --link-dest / --copy-dest / --compare-dest
                         candidate resolution + its mtime-window helper
  backup.rs            - --backup staging and forced destination removal
                         + the two stale-backup helpers only it calls
  deferred_updates.rs  - --delay-updates recording and commit

state.rs keeps construction, timeouts, caches and error accounting and
drops to 554 lines; no file exceeds 464.

Code movement only: the function set is identical before and after (56
in both), bodies are unchanged, and every upstream citation is carried
across verbatim. The only rewrite is new(), whose five inline
derivations become five named calls.

Verified: fmt, workspace clippy --all-targets --all-features, engine
4973/4973, and the local --write-batch round-trip.
