+++ crates/engine/src/local_copy/context_impl/state.rs
@@ -30,137 +30,11 @@ impl<'a> CopyContext<'a> {
             DeferredSync::new(SyncStrategy::None)
         };
 
-        // When batch mode is active, create a temp file to buffer per-file
-        // delta data. The flist entries go directly to the batch writer during
-        // the walk, but file data (NDX + iflags + sum_head + tokens + checksum)
-        // must come AFTER the flist end marker in the batch stream.
-        let batch_delta_buf = options
-            .get_batch_writer()
-            .map(|_| std::io::Cursor::new(Vec::new()));
-
-        let batch_ndx_codec = options.get_batch_writer().map(|batch_writer_arc| {
-            let guard = batch_writer_arc.lock().expect("batch writer mutex poisoned");
-            let proto_version = guard.config().protocol_version;
-            drop(guard);
-            protocol::codec::NdxCodecEnum::new(proto_version as u8)
-        });
-
-        let batch_flist_writer = options.get_batch_writer().map(|batch_writer_arc| {
-            let guard = batch_writer_arc.lock().expect("batch writer mutex poisoned");
-            let proto_version = guard.config().protocol_version;
-            let compat_flags_val = guard.config().compat_flags;
-            let checksum_seed = guard.config().checksum_seed;
-            // The preserve flags come from the header this same writer already
-            // recorded, not from `options`, because the header is what the
-            // reader configures itself from (batch/src/reader/flist.rs). Unlike
-            // upstream - whose batch file is a byte tee of one real stream, so
-            // a writer/reader disagreement is unspellable - the local
-            // --write-batch path re-encodes the flist with a second encoder,
-            // and deriving it from anything but the recorded flags lets the two
-            // sides drift into an undecodable batch.
-            let stream_flags = *guard.stream_flags();
-            drop(guard);
-            let protocol = protocol::ProtocolVersion::try_from(proto_version as u8)
-                .unwrap_or(protocol::ProtocolVersion::NEWEST);
-            // upstream: io.c:start_write_batch() - compat_flags are written to the batch
-            // header. The flist writer must use the same compat_flags to ensure the wire
-            // encoding (varint flags, safe file list) matches what the reader will expect
-            // when decoding the batch body.
-            let writer = if let Some(cf) = compat_flags_val {
-                let compat = protocol::CompatibilityFlags::from_bits(cf as u32);
-                protocol::flist::FileListWriter::with_compat_flags(protocol, compat)
-            } else {
-                protocol::flist::FileListWriter::new(protocol)
-            };
-            let writer = writer
-                .with_preserve_uid(stream_flags.preserve_uid)
-                .with_preserve_gid(stream_flags.preserve_gid)
-                .with_preserve_links(stream_flags.preserve_links)
-                .with_preserve_devices(stream_flags.preserve_devices)
-                // upstream: batch.c flag_ptr[] bit 4 is `preserve_devices` and
-                // covers --specials too (upstream `-D` = both), so the reader
-                // derives specials from that one bit. Deriving it here from
-                // `options.specials_enabled()` instead would desync a
-                // `--specials --no-devices` batch.
-                .with_preserve_specials(stream_flags.preserve_devices)
-                .with_preserve_hard_links(stream_flags.preserve_hard_links)
-                .with_preserve_acls(stream_flags.preserve_acls)
-                .with_preserve_xattrs(stream_flags.preserve_xattrs)
-                // --atimes / --crtimes have no flag_ptr[] bit; the reader takes
-                // them from the replay invocation instead (reader/flist.rs), so
-                // the writer keeps using the live options here.
-                .with_preserve_atimes(options.preserve_atimes())
-                .with_preserve_crtimes(options.preserve_crtimes())
-                // The batch id-list trailer (write_batch_id_lists) emits bare
-                // terminators without names, so owner names must ride inline in
-                // every entry regardless of inc_recurse. Keep XMIT_*_NAME_FOLLOWS
-                // always enabled for the batch flist writer.
-                .with_name_follows(true);
-
-            // upstream: flist.c:162 - under always_checksum every regular-file
-            // entry carries a trailing digest. The reader re-derives the length
-            // from the protocol version (batch/src/reader/flist.rs
-            // default_flist_csum_len), so the two must agree or the reader
-            // walks off the end of each entry.
-            if stream_flags.always_checksum {
-                writer
-                    .with_always_checksum(batch::reader::default_flist_csum_len(proto_version))
-                    .with_checksum_seed(checksum_seed)
-            } else {
-                writer
-            }
-        });
-
-        // upstream: batch.c:68 - stream-flag bit 8 records `do_compression`, and
-        // the batch file is a tee of the wire stream, so a batch written under
-        // `-z` carries `token.c:send_deflated_token()` framing. The codec is
-        // always zlib: `compat.c:414 getenv_nstr()` forces the compression list
-        // to "zlib" whenever `write_batch` is set ("When writing a batch file,
-        // we always negotiate an old-style choice"), which is also what
-        // `compat.c:194-195 parse_compress_choice()` assumes on `--read-batch`.
-        let batch_token_encoder = options.get_batch_writer().and_then(|batch_writer_arc| {
-            if !options.compress_enabled() {
-                return None;
-            }
-            let proto_version = batch_writer_arc
-                .lock()
-                .expect("batch writer mutex poisoned")
-                .config()
-                .protocol_version;
-            Some(protocol::wire::CompressedTokenEncoder::new(
-                options.compression_level(),
-                proto_version as u32,
-            ))
-        });
-
-        let adaptive_level = if options.compress_enabled() && options.adaptive_compress_enabled() {
-            let initial = options.compression_level();
-            let level_i32 = match initial {
-                CompressionLevel::None => 0,
-                CompressionLevel::Fast => 1,
-                CompressionLevel::Default => 6,
-                CompressionLevel::Best => 9,
-                CompressionLevel::Precise(v) => i32::from(v.get()),
-                // upstream: token.c:73 - preserve zstd's negative "fast" levels.
-                CompressionLevel::PreciseSigned(v) => v,
-            };
-            match options.compression_algorithm() {
-                CompressionAlgorithm::Zlib => Some(
-                    compress::strategy::adaptive_level::AdaptiveLevelController::for_zlib(
-                        level_i32,
-                    ),
-                ),
-                #[cfg(feature = "zstd")]
-                CompressionAlgorithm::Zstd => Some(
-                    compress::strategy::adaptive_level::AdaptiveLevelController::for_zstd(
-                        level_i32,
-                    ),
-                ),
-                _ => None,
-            }
-        } else {
-            None
-        };
+        let batch_delta_buf = build_batch_delta_buffer(&options);
+        let batch_ndx_codec = build_batch_ndx_codec(&options);
+        let batch_flist_writer = build_batch_flist_writer(&options);
+        let batch_token_encoder = build_batch_token_encoder(&options);
+        let adaptive_level = build_adaptive_level_controller(&options);
 
         Self {
             mode,
@@ -478,256 +352,6 @@ impl<'a> CopyContext<'a> {
         Ok(())
     }
 
-    /// Searches `--link-dest` directories for the BEST file matching the source.
-    ///
-    /// upstream: generator.c:954-983 `try_dests_reg()` scans every basis dir and
-    /// tracks the highest match_level (2 = data matches `quick_check_ok`, 3 = data
-    /// and attributes both match `unchanged_attrs`), breaking early only on an
-    /// exact (level-3) match. Returning the first data-only candidate instead
-    /// would let an earlier match_level-2 basis (attrs differ) shadow a later
-    /// exact one, forcing an unnecessary copy + attr reapply where upstream would
-    /// hard-link the exact basis with no reapply. The caller re-derives the
-    /// winning candidate's level to choose hard-link vs copy, so returning the
-    /// best candidate is sufficient to mirror upstream.
-    pub(super) fn link_dest_target(
-        &self,
-        relative: &Path,
-        source: &Path,
-        metadata: &fs::Metadata,
-        size_only: bool,
-        ignore_times: bool,
-        checksum: bool,
-    ) -> Result<Option<PathBuf>, LocalCopyError> {
-        if self.options.link_dest_entries().is_empty() {
-            return Ok(None);
-        }
-
-        let metadata_options = self.metadata_options();
-        let preserve_xattrs = {
-            #[cfg(all(unix, feature = "xattr"))]
-            {
-                self.options.preserve_xattrs()
-            }
-            #[cfg(not(all(unix, feature = "xattr")))]
-            {
-                false
-            }
-        };
-
-        let mut best: Option<(PathBuf, u8)> = None;
-        for entry in self.options.link_dest_entries() {
-            let candidate = entry.resolve(self.destination_root(), relative);
-            let candidate_metadata = match fs::metadata(&candidate) {
-                Ok(metadata) => metadata,
-                Err(error) if error.kind() == io::ErrorKind::NotFound => continue,
-                Err(error) => {
-                    return Err(LocalCopyError::io(
-                        "inspect link-dest candidate",
-                        candidate,
-                        error,
-                    ));
-                }
-            };
-
-            if !candidate_metadata.file_type().is_file() {
-                continue;
-            }
-
-            if !should_skip_copy(CopyComparison {
-                source_path: source,
-                source: metadata,
-                destination_path: candidate.as_path(),
-                destination: &candidate_metadata,
-                size_only,
-                ignore_times,
-                checksum,
-                checksum_algorithm: self.options.checksum_algorithm(),
-                modify_window: self.options.modify_window(),
-                prefetched_match: None,
-            }) {
-                continue;
-            }
-
-            // At least match_level 2 (data matches); match_level 3 when the
-            // preserved attributes also match, which the caller hard-links.
-            let level = if crate::local_copy::reference_attrs_unchanged(
-                &candidate,
-                source,
-                metadata,
-                &metadata_options,
-                preserve_xattrs,
-            ) {
-                3
-            } else {
-                2
-            };
-
-            if best.as_ref().is_none_or(|(_, best_level)| level > *best_level) {
-                best = Some((candidate, level));
-            }
-            // upstream: generator.c:979 - an exact match ends the scan.
-            if level == 3 {
-                break;
-            }
-        }
-
-        Ok(best.map(|(candidate, _)| candidate))
-    }
-
-    /// Locates a `--link-dest` basis symlink at `relative` that points at the
-    /// same `target`.
-    ///
-    /// Returns the basis symlink path when a link-dest entry holds a symlink
-    /// with a matching target, so the receiver can hard-link the symlink into
-    /// place (`hL`) instead of recreating it.
-    ///
-    /// upstream: generator.c:1117-1134 try_dests_non() - LINK_DEST hard-links a
-    /// matching symlink from the basis when CAN_HARDLINK_SYMLINK is supported.
-    pub(super) fn link_dest_symlink_target(
-        &self,
-        relative: &Path,
-        target: &Path,
-    ) -> Result<Option<PathBuf>, LocalCopyError> {
-        if self.options.link_dest_entries().is_empty() {
-            return Ok(None);
-        }
-
-        for entry in self.options.link_dest_entries() {
-            let candidate = entry.resolve(self.destination_root(), relative);
-            let candidate_metadata = match fs::symlink_metadata(&candidate) {
-                Ok(metadata) => metadata,
-                Err(error) if error.kind() == io::ErrorKind::NotFound => continue,
-                Err(error) => {
-                    return Err(LocalCopyError::io(
-                        "inspect link-dest symlink",
-                        candidate,
-                        error,
-                    ));
-                }
-            };
-
-            if !candidate_metadata.file_type().is_symlink() {
-                continue;
-            }
-
-            match fs::read_link(&candidate) {
-                Ok(basis_target) if basis_target == target => return Ok(Some(candidate)),
-                Ok(_) => continue,
-                Err(error) => {
-                    return Err(LocalCopyError::io(
-                        "read link-dest symlink",
-                        candidate,
-                        error,
-                    ));
-                }
-            }
-        }
-
-        Ok(None)
-    }
-
-    /// Locates a `--link-dest` basis device or special file at `relative` that
-    /// exactly matches the source node, returning its path so the receiver can
-    /// hard-link it into place (`hD`/`hS` + blank) instead of recreating it.
-    ///
-    /// Mirrors upstream `generator.c:1064-1152` try_dests_non(): a `LINK_DEST`
-    /// basis entry of the same file-type bucket (`FT_DEVICE`/`FT_SPECIAL`) whose
-    /// device number (devices) or `_S_IFMT` (specials) matches
-    /// (`generator.c:664-678` quick_check_ok) AND whose preserved attributes are
-    /// unchanged (`generator.c:468-507` unchanged_attrs) reaches match_level 3
-    /// and is hard-linked, itemizing as an exact match. `CAN_HARDLINK_SPECIAL`
-    /// is defined on Linux, so devices and specials participate.
-    #[cfg(unix)]
-    pub(super) fn link_dest_special_target(
-        &self,
-        relative: &Path,
-        metadata: &fs::Metadata,
-        metadata_options: &MetadataOptions,
-    ) -> Result<Option<PathBuf>, LocalCopyError> {
-        use std::os::unix::fs::{FileTypeExt, MetadataExt};
-
-        if self.options.link_dest_entries().is_empty() {
-            return Ok(None);
-        }
-
-        let source_type = metadata.file_type();
-        let source_is_device = source_type.is_block_device() || source_type.is_char_device();
-        let source_is_special = source_type.is_fifo() || source_type.is_socket();
-        if !source_is_device && !source_is_special {
-            return Ok(None);
-        }
-
-        let modify_window = self.options.modify_window();
-
-        for entry in self.options.link_dest_entries() {
-            let candidate = entry.resolve(self.destination_root(), relative);
-            let candidate_metadata = match fs::symlink_metadata(&candidate) {
-                Ok(metadata) => metadata,
-                Err(error) if error.kind() == io::ErrorKind::NotFound => continue,
-                Err(error) => {
-                    return Err(LocalCopyError::io(
-                        "inspect link-dest special",
-                        candidate,
-                        error,
-                    ));
-                }
-            };
-
-            let cand_type = candidate_metadata.file_type();
-
-            // upstream: generator.c:1076 - the basis must share the source's
-            // file-type bucket, and generator.c:657-671 quick_check_ok compares
-            // st_rdev (devices) or _S_IFMT (specials, i.e. fifo vs socket).
-            if source_is_device {
-                if !(cand_type.is_block_device() || cand_type.is_char_device()) {
-                    continue;
-                }
-                if metadata.rdev() != candidate_metadata.rdev() {
-                    continue;
-                }
-            } else if source_type.is_fifo() != cand_type.is_fifo()
-                || source_type.is_socket() != cand_type.is_socket()
-            {
-                continue;
-            }
-
-            // upstream: generator.c:468-507 unchanged_attrs - preserved mtime,
-            // perms and ownership must match for the match_level-3 hard-link.
-            if metadata_options.times()
-                && !mtimes_within_window(metadata, &candidate_metadata, modify_window)
-            {
-                continue;
-            }
-            if metadata_options.permissions()
-                && (metadata.mode() & 0o7777) != (candidate_metadata.mode() & 0o7777)
-            {
-                continue;
-            }
-            if metadata_options.owner() && metadata.uid() != candidate_metadata.uid() {
-                continue;
-            }
-            if metadata_options.group() && metadata.gid() != candidate_metadata.gid() {
-                continue;
-            }
-
-            return Ok(Some(candidate));
-        }
-
-        Ok(None)
-    }
-
-    /// Non-Unix stub: device and special nodes cannot be materialised, so no
-    /// `--link-dest` basis can ever match one.
-    #[cfg(not(unix))]
-    pub(super) fn link_dest_special_target(
-        &self,
-        _relative: &Path,
-        _metadata: &fs::Metadata,
-        _metadata_options: &MetadataOptions,
-    ) -> Result<Option<PathBuf>, LocalCopyError> {
-        Ok(None)
-    }
-
     /// Returns the configured `--link-dest` / `--copy-dest` / `--compare-dest`
     /// reference directories.
     pub(super) fn reference_directories(&self) -> &[ReferenceDirectory] {
@@ -786,596 +410,6 @@ impl<'a> CopyContext<'a> {
         self.destination_metadata_cache.clear();
     }
 
-    /// Records a finalized directory for the final
-    /// [`touch_up_dirs`](Self::touch_up_dirs) pass.
-    ///
-    /// Called by `apply_final_directory_metadata` during the traversal. Late
-    /// in-directory mutations (the delayed-update rename sweep, deferred
-    /// deletions, backup file creation) bump the directory mtime and require the
-    /// directory to stay writable, so the final pass re-applies the recorded
-    /// source mtime and reinstates the restricted mode last.
-    ///
-    /// `mtime` is the source mtime to re-apply (when times are preserved) and
-    /// `restore_mode` is the restricted mode to reinstate for a directory that
-    /// was kept writable during the transfer.
-    pub(super) fn record_finalized_directory(
-        &mut self,
-        destination: &Path,
-        mtime: Option<filetime::FileTime>,
-        restore_mode: Option<u32>,
-    ) {
-        self.deferred_ops.finalized_dirs.push(FinalizedDir {
-            path: destination.to_path_buf(),
-            mtime,
-            restore_mode,
-        });
-    }
-
-    /// Queues a deferred update for `--delay-updates` and records the hard-link
-    /// source. The staging directory is tracked for cleanup after commit.
-    pub(super) fn register_deferred_update(&mut self, update: DeferredUpdate) {
-        // Track the `.~tmp~` staging directory for cleanup after all updates
-        // are committed.
-        if let Some(parent) = update.guard.staging_path().parent() {
-            if parent
-                .file_name()
-                .is_some_and(|name| name == super::options::staging::DELAY_UPDATES_PARTIAL_DIR)
-            {
-                self.deferred_ops
-                    .delay_staging_dirs
-                    .insert(parent.to_path_buf());
-            }
-        }
-        let metadata = update.metadata.clone();
-        let destination = update.destination.clone();
-        self.record_hard_link(&metadata, destination.as_path());
-        self.deferred_ops.updates.push(update);
-    }
-
-    /// Commits a single deferred update matching the given destination path,
-    /// if one exists in the queue.
-    pub(super) fn commit_deferred_update_for(
-        &mut self,
-        destination: &Path,
-    ) -> Result<(), LocalCopyError> {
-        if let Some(index) = self
-            .deferred_ops
-            .updates
-            .iter()
-            .position(|update| update.destination.as_path() == destination)
-        {
-            let update = self.deferred_ops.updates.swap_remove(index);
-            self.finalize_deferred_update(update)?;
-        }
-        Ok(())
-    }
-
-    /// Commits all remaining deferred updates and removes empty staging
-    /// directories.
-    pub(super) fn flush_deferred_updates(&mut self) -> Result<(), LocalCopyError> {
-        let updates = std::mem::take(&mut self.deferred_ops.updates);
-        for update in updates {
-            self.finalize_deferred_update(update)?;
-        }
-
-        // Remove empty `.~tmp~` staging directories after all deferred files
-        // have been moved to their final locations. This covers both updates
-        // committed here and those committed early via `commit_deferred_update_for`.
-        //
-        // upstream: receiver.c -- handle_partial_dir(partialptr, PDIR_DELETE)
-        let dirs = std::mem::take(&mut self.deferred_ops.delay_staging_dirs);
-        for dir in &dirs {
-            let _ = fs::remove_dir(dir);
-        }
-
-        Ok(())
-    }
-
-    /// Hard-links, renames, or copies an existing destination entry to the
-    /// backup location when `--backup` is enabled.
-    ///
-    /// `prefer_rename` mirrors upstream's `make_backup(fname, prefer_rename)`
-    /// parameter: callers backing up an item that is about to be unlinked
-    /// outright (the delete pass) pass `true` to skip straight to the rename
-    /// tier, since the caller removes the original right after regardless of
-    /// which strategy placed the backup (`delete.c:165-167`). Callers backing
-    /// up an item before overwriting it with fresh content pass `false` so
-    /// the hard-link tier runs first (`rsync.c:740`, `receiver.c:538`).
-    ///
-    /// Emits an `--info=BACKUP` notice mirroring upstream rsync 3.4.1
-    /// (backup.c:352) under `INFO_GTE(BACKUP, 1)` once the backup has been
-    /// placed successfully. The wording matches upstream byte-for-byte:
-    /// `backed up <fname> to <buf>`.
-    /// Reports whether `--safe-links` forbids backing up `destination`,
-    /// emitting upstream's `SYMSAFE,1` notice when it does.
-    ///
-    /// upstream: `backup.c:289-310` gates the hard-link/rename fast path and
-    /// `backup.c:368-375` gates the copy fallback on the same rule. Upstream
-    /// carries both because `link_or_rename()` would otherwise hard-link an
-    /// escaping symlink into the backup area and `goto success`, never reaching
-    /// the copy path's check - its own comment at `backup.c:282-288` says so.
-    /// Routing both oc sites through one predicate is what stops them drifting
-    /// apart again, which is exactly how the fast path came to be unguarded.
-    ///
-    /// A `readlink` failure refuses the backup rather than permitting it:
-    /// upstream fails closed at `backup.c:295-300` on the same reasoning, that
-    /// an unverifiable target must not be preserved into the backup area.
-    fn backup_refused_by_safe_links(&self, destination: &Path, file_type: fs::FileType) -> bool {
-        if !(file_type.is_symlink()
-            && self.options.links_enabled()
-            && self.options.safe_links_enabled())
-        {
-            return false;
-        }
-
-        let Ok(target) = fs::read_link(destination) else {
-            // upstream: backup.c:296-297
-            info_log!(
-                Symsafe,
-                1,
-                "not backing up symlink with unreadable target \"{}\"",
-                destination.display()
-            );
-            return true;
-        };
-
-        let safety_rel = destination
-            .strip_prefix(self.destination_root())
-            .unwrap_or(destination);
-        if symlink_target_is_safe(&target, safety_rel) {
-            return false;
-        }
-
-        // upstream: backup.c:303-306
-        info_log!(
-            Symsafe,
-            1,
-            "not backing up unsafe symlink \"{}\" -> \"{}\"",
-            destination.display(),
-            target.display()
-        );
-        true
-    }
-
-    pub(super) fn backup_existing_entry(
-        &mut self,
-        destination: &Path,
-        _relative: Option<&Path>,
-        file_type: fs::FileType,
-        prefer_rename: bool,
-    ) -> Result<(), LocalCopyError> {
-        if !self.options.backup_enabled() || self.mode.is_dry_run() {
-            return Ok(());
-        }
-
-        if file_type.is_dir() {
-            return Ok(());
-        }
-
-        // Always derive the relative path from destination/destination_root
-        // rather than using the source-relative path. The source-relative path
-        // may not include the source directory basename (e.g., "nested/file.txt"
-        // instead of "source/nested/file.txt"), causing backup files to be
-        // placed at the wrong location when --backup-dir is used.
-        // upstream: backup.c:get_backup_name() uses the destination-relative path
-        let backup_path = compute_backup_path(
-            self.destination_root(),
-            destination,
-            None,
-            self.options.backup_directory(),
-            self.options.backup_suffix(),
-        );
-
-        if let Some(parent) = backup_path.parent()
-            && !parent.as_os_str().is_empty()
-        {
-            // upstream: backup.c:copy_valid_path() - create any new backup
-            // subdirectories, mirroring the source dir attrs onto each and
-            // clearing non-directory obstructions.
-            create_backup_parents(
-                self.destination_root(),
-                self.options.backup_directory(),
-                parent,
-                &self.metadata_options(),
-            )?;
-        }
-
-        // upstream: backup.c:289-310 - honour --safe-links BEFORE the
-        // hard-link/rename fast path. Without this the fast path preserves an
-        // escaping symlink in the backup area and returns successfully, so the
-        // cross-device check further down is never consulted.
-        if self.backup_refused_by_safe_links(destination, file_type) {
-            return Ok(());
-        }
-
-        // upstream: backup.c:200-207 link_or_rename() - try a hard link into
-        // the backup area first when the caller doesn't prefer a rename
-        // outright. A successful link leaves the destination's inode as the
-        // backup, so no metadata reapply is needed below (same as RENAME).
-        let hard_linked = if prefer_rename {
-            None
-        } else {
-            match create_hard_link(destination, &backup_path) {
-                Ok(()) => Some(BackupStrategy::HardLink),
-                Err(error) if error.kind() == io::ErrorKind::NotFound => return Ok(()),
-                // upstream: backup.c:247-256 - a stale backup entry is
-                // removed and the hard link retried once.
-                Err(error) if is_stale_backup_conflict(&error) => {
-                    remove_stale_backup_entry(&backup_path)?;
-                    match create_hard_link(destination, &backup_path) {
-                        Ok(()) => Some(BackupStrategy::HardLink),
-                        Err(_) => None,
-                    }
-                }
-                // Any other failure (e.g. EXDEV across a `--backup-dir` on a
-                // different filesystem, or a type the platform cannot
-                // hard-link) falls through to the rename tier below, mirroring
-                // `link_or_rename`'s in-call rename fallback.
-                Err(_) => None,
-            }
-        };
-
-        // Track which backup strategy succeeded so we can emit the matching
-        // upstream `--debug=BACKUP` trace (HLINK, RENAME, COPY, SYMLINK, or
-        // DEVICE).
-        // upstream: backup.c:link_or_rename and the fall-through copy_file path.
-        let strategy = if let Some(strategy) = hard_linked {
-            strategy
-        } else {
-            match backup_rename(destination, &backup_path) {
-            Ok(()) => BackupStrategy::Rename,
-            Err(error) if error.kind() == io::ErrorKind::NotFound => return Ok(()),
-            // upstream: backup.c:247-256 - link_or_rename failing with EEXIST or
-            // EISDIR is recoverable: lstat the target and call delete_item with
-            // DEL_RECURSE, then retry. EISDIR fires when the backup-dir already
-            // contains a directory at the path we need (e.g. user pre-created
-            // it, or a previous backup left a tree there); without this arm,
-            // backup test 4 from upstream backup.test fails as exit-23 fatal.
-            // Windows reports renaming a file onto an existing directory as
-            // ERROR_ACCESS_DENIED (PermissionDenied) rather than EEXIST/EISDIR,
-            // so it must enter the same recovery arm; the inner symlink_metadata
-            // re-stat below still gates removal on the target actually existing,
-            // so a genuine permission error falls through to the retry-and-fail
-            // path unchanged.
-            Err(error) if is_stale_backup_conflict(&error) => {
-                remove_stale_backup_entry(&backup_path)?;
-                fs::rename(destination, &backup_path).map_err(|rename_error| {
-                    LocalCopyError::io("create backup", backup_path.clone(), rename_error)
-                })?;
-                BackupStrategy::Rename
-            }
-            Err(error) if error.kind() == io::ErrorKind::CrossesDevices => {
-                // upstream: backup.c:368-375 - the copy fallback re-checks
-                // --safe-links before recreating the symlink. Upstream keeps
-                // this alongside the fast-path check above; both go through the
-                // one predicate so they cannot implement different rules.
-                if self.backup_refused_by_safe_links(destination, file_type) {
-                    return Ok(());
-                }
-                match copy_entry_to_backup(
-                    destination,
-                    &backup_path,
-                    file_type,
-                    self.options.devices_enabled(),
-                    self.options.specials_enabled(),
-                    self.options.fake_super_enabled(),
-                )? {
-                    Some(strategy) => strategy,
-                    // upstream: backup.c:306-317 - a non-regular file that is
-                    // neither backed up as a device/special (gates off) nor a
-                    // symlink is skipped; make_backup returns 3 and leaves no
-                    // backup, so emit no trace and no "backed up" notice.
-                    None => return Ok(()),
-                }
-            }
-            Err(error) => {
-                return Err(LocalCopyError::io("create backup", backup_path, error));
-            }
-            }
-        };
-
-        // upstream: backup.c:338-341 - set_file_attrs(buf, file, NULL, fname,
-        // ATTRS_ACCURATE_TIME) copies the source node's mode/owner/times onto
-        // the freshly-created backup node (with preserve_xattrs temporarily
-        // cleared), overriding the umask/copy defaults. Every cross-device copy
-        // fallback needs this - regular file (COPY), symlink (SYMLINK), and
-        // device/special (DEVICE) - because fs::copy/create_symlink/mknod leave
-        // the backup owned by the caller (root), whereas upstream reapplies the
-        // source uid/gid/mode/mtime. The same-fs rename path carries the inode's
-        // attributes verbatim and is skipped here.
-        if matches!(
-            strategy,
-            BackupStrategy::Copy | BackupStrategy::Symlink | BackupStrategy::Device
-        ) && let Ok(source_meta) = fs::symlink_metadata(destination)
-        {
-            let metadata_options = self.metadata_options();
-            // upstream: rsync.c:set_file_attrs() skips chmod for symlinks and
-            // applies ownership/times with AT_SYMLINK_NOFOLLOW.
-            if strategy == BackupStrategy::Symlink {
-                apply_symlink_metadata_with_options(&backup_path, &source_meta, &metadata_options)
-                    .map_err(map_metadata_error)?;
-            } else {
-                apply_file_metadata_with_options(&backup_path, &source_meta, &metadata_options)
-                    .map_err(map_metadata_error)?;
-            }
-            // upstream: xattrs.c:set_stat_xattr() re-records the source stat in
-            // `user.rsync.%stat` under --fake-super so the virtualised node's
-            // mode/owner/rdev survive; no-op when fake-super is off.
-            #[cfg(all(unix, feature = "xattr"))]
-            store_effective_fake_super_if_requested(
-                &metadata_options,
-                destination,
-                &backup_path,
-                &source_meta,
-            )?;
-        }
-
-        // upstream: backup.c:201-202,216-217,282-283,299-300,333-334 -
-        // DEBUG_GTE(BACKUP, 1) emits one of HLINK/RENAME/DEVICE/SYMLINK/COPY
-        // per success path. oc-rsync's local-copy executor prefers a
-        // same-filesystem hard link, falls back to rename, then falls back to
-        // copy or symlink recreation across filesystem boundaries.
-        let destination_display = destination.display().to_string();
-        match strategy {
-            BackupStrategy::HardLink => trace_make_backup_hlink(&destination_display),
-            BackupStrategy::Rename => trace_make_backup_rename(&destination_display),
-            BackupStrategy::Copy => trace_make_backup_copy(&destination_display),
-            BackupStrategy::Symlink => trace_make_backup_symlink(&destination_display),
-            BackupStrategy::Device => trace_make_backup_device(&destination_display),
-        }
-
-        // upstream: backup.c:433 - rprintf(FINFO, "backed up %s to %s\n", fname, buf)
-        // emits fname and buf as the rsync-relative paths (e.g. "deep/name1"),
-        // not absolute filesystem paths. Strip the destination_root prefix so
-        // the message matches upstream byte-for-byte and grep-by-relative-path
-        // assertions in the upstream backup.test pass.
-        let dest_root = self.destination_root();
-        let destination_rel = destination.strip_prefix(dest_root).unwrap_or(destination);
-        let backup_rel = backup_path
-            .strip_prefix(dest_root)
-            .unwrap_or(backup_path.as_path());
-        info_log!(
-            Backup,
-            1,
-            "backed up {} to {}",
-            destination_rel.display(),
-            backup_rel.display()
-        );
-
-        Ok(())
-    }
-
-    /// Backs up an existing regular destination file by COPYING it aside,
-    /// leaving the destination inode in place for an `--inplace` rewrite.
-    ///
-    /// Unlike [`backup_existing_entry`](Self::backup_existing_entry), which
-    /// renames the destination (moving its inode to the backup name), this
-    /// duplicates the pre-transfer contents so the original inode stays put and
-    /// is updated in place - the whole point of `--inplace`. Only used for
-    /// regular-file inplace updates; other entries keep the rename path.
-    ///
-    /// upstream: backup.c make_backup() inplace copy path - the generator makes
-    /// the backup a COPY (`generator.c:1862` `copy_file(fname, backupptr, ...)`,
-    /// with the delta twin building the same copy at `generator.c:1898`) BEFORE
-    /// the receiver rewrites the destination in place. The inplace copy bypasses
-    /// `make_backup()`, so it emits no `DEBUG_GTE(BACKUP, 1)` trace, but it still
-    /// emits the `INFO_GTE(BACKUP, 1)` "backed up X to Y" line
-    /// (`generator.c:1990-1992`).
-    pub(super) fn backup_existing_entry_copy(
-        &mut self,
-        destination: &Path,
-    ) -> Result<(), LocalCopyError> {
-        let backup_path = compute_backup_path(
-            self.destination_root(),
-            destination,
-            None,
-            self.options.backup_directory(),
-            self.options.backup_suffix(),
-        );
-
-        if let Some(parent) = backup_path.parent()
-            && !parent.as_os_str().is_empty()
-        {
-            // upstream: backup.c:copy_valid_path() - create any new backup
-            // subdirectories, mirroring the source dir attrs onto each and
-            // clearing non-directory obstructions.
-            create_backup_parents(
-                self.destination_root(),
-                self.options.backup_directory(),
-                parent,
-                &self.metadata_options(),
-            )?;
-        }
-
-        // upstream: generator.c:1866 copy_file() - duplicate the pre-image; the
-        // destination inode is left in place to be rewritten by the inplace
-        // writer. A pre-existing backup is overwritten (upstream robust_unlinks
-        // it at generator.c:1901); fs::copy truncates to the same end state.
-        match fs::copy(destination, &backup_path) {
-            Ok(_) => {}
-            // A vanished destination has nothing to back up, mirroring the
-            // NotFound arm of the rename path (backup.c:make_backup returns 3).
-            Err(error) if error.kind() == io::ErrorKind::NotFound => return Ok(()),
-            Err(error) => {
-                return Err(LocalCopyError::io("create backup", backup_path, error));
-            }
-        }
-
-        // upstream: generator.c:1988 set_file_attrs(backupptr, back_file, ...) -
-        // the backup carries the source node's mode/owner/times. fs::copy
-        // preserves only permissions, so reapply full metadata like the
-        // cross-device COPY fallback in backup_existing_entry.
-        if let Ok(source_meta) = fs::symlink_metadata(destination) {
-            let metadata_options = self.metadata_options();
-            apply_file_metadata_with_options(&backup_path, &source_meta, &metadata_options)
-                .map_err(map_metadata_error)?;
-            // upstream: generator.c:1985 copy_xattrs() under --fake-super
-            // re-records the source stat in `user.rsync.%stat`; no-op otherwise.
-            #[cfg(all(unix, feature = "xattr"))]
-            store_effective_fake_super_if_requested(
-                &metadata_options,
-                destination,
-                &backup_path,
-                &source_meta,
-            )?;
-        }
-
-        // upstream: generator.c:1990-1992 - INFO_GTE(BACKUP, 1) "backed up X to
-        // Y". No DEBUG_GTE(BACKUP) trace: the inplace copy bypasses make_backup().
-        let dest_root = self.destination_root();
-        let destination_rel = destination.strip_prefix(dest_root).unwrap_or(destination);
-        let backup_rel = backup_path
-            .strip_prefix(dest_root)
-            .unwrap_or(backup_path.as_path());
-        info_log!(
-            Backup,
-            1,
-            "backed up {} to {}",
-            destination_rel.display(),
-            backup_rel.display()
-        );
-
-        Ok(())
-    }
-
-    /// Forcibly removes a type-conflicting destination entry (backing it up
-    /// first if needed) to make room for an incoming item of a different type.
-    ///
-    /// upstream: generator.c:1240 recv_generator() clears the conflicting
-    /// destination with `delete_item(fname, mode, del_opts | DEL_FOR_FILE)`.
-    /// delete.c:178-194 `delete_item()` suppresses `log_delete()` and the
-    /// `stats.deleted_files` bump whenever `flags & DEL_MAKE_ROOM` is set, so
-    /// the make-room removal of the conflicting entry itself is silent and
-    /// uncounted (unlike a genuine delete-pass deletion). When that entry is a
-    /// directory, `delete_dir_contents()` (delete.c:83) recurses with
-    /// DEL_MAKE_ROOM stripped, so its contents are itemized (`*deleting`) and
-    /// counted like ordinary deletions while the directory node stays silent.
-    pub(super) fn force_remove_destination(
-        &mut self,
-        destination: &Path,
-        relative: Option<&Path>,
-        metadata: &fs::Metadata,
-    ) -> Result<(), LocalCopyError> {
-        let file_type = metadata.file_type();
-
-        if self.mode.is_dry_run() {
-            if file_type.is_dir() {
-                self.record_make_room_contents(destination, relative)?;
-            }
-            self.register_progress();
-            return Ok(());
-        }
-
-        // upstream: delete.c:165-167 - `delete_item()` (the callee behind
-        // generator.c:1240's DEL_FOR_FILE removal) always calls
-        // `make_backup(fbuf, True)`, so the hard-link tier is skipped here
-        // exactly as it is for a genuine delete-pass removal.
-        self.backup_existing_entry(destination, relative, file_type, true)?;
-
-        if file_type.is_dir() {
-            self.record_make_room_contents(destination, relative)?;
-        }
-
-        let context = if file_type.is_dir() {
-            "remove existing directory"
-        } else {
-            "remove existing destination"
-        };
-
-        let removal_result = if file_type.is_dir() {
-            fs::remove_dir_all(destination)
-        } else {
-            fs::remove_file(destination)
-        };
-
-        match removal_result {
-            Ok(()) => {}
-            Err(error) if error.kind() == io::ErrorKind::NotFound => {}
-            Err(error) => {
-                return Err(LocalCopyError::io(
-                    context,
-                    destination.to_path_buf(),
-                    error,
-                ));
-            }
-        }
-
-        self.register_progress();
-
-        Ok(())
-    }
-
-    /// Itemizes and counts the contents of a conflicting directory that is
-    /// being cleared to make room for an incoming item, mirroring upstream's
-    /// `delete_dir_contents()` recursion (delete.c:83): the children are
-    /// reported like delete-pass deletions (DEL_MAKE_ROOM stripped) while the
-    /// directory node itself is removed silently by the caller.
-    fn record_make_room_contents(
-        &mut self,
-        destination: &Path,
-        relative: Option<&Path>,
-    ) -> Result<(), LocalCopyError> {
-        let mut subtree_path = destination.to_path_buf();
-        let mut subtree_relative = relative.map(Path::to_path_buf).unwrap_or_else(|| {
-            destination
-                .file_name()
-                .map(PathBuf::from)
-                .unwrap_or_default()
-        });
-        record_directory_subtree(self, &mut subtree_path, &mut subtree_relative)
-    }
-
-    /// Commits a single deferred update: moves the staged file to its final
-    /// path and applies metadata.
-    pub(super) fn finalize_deferred_update(
-        &mut self,
-        update: DeferredUpdate,
-    ) -> Result<(), LocalCopyError> {
-        let DeferredUpdate {
-            guard,
-            metadata,
-            metadata_options,
-            mode,
-            path_context,
-            destination,
-            #[cfg(all(unix, feature = "xattr"))]
-            preserve_xattrs,
-            #[cfg(all(any(unix, windows), feature = "acl"))]
-            preserve_acls,
-        } = update;
-
-        #[cfg(not(any(all(unix, feature = "xattr"), all(any(unix, windows), feature = "acl"))))]
-        let _ = &path_context.source;
-
-        // Deferred updates have no open fd, so the cross-device flag is unused.
-        let _cross_device = guard.commit()?;
-
-        self.apply_metadata_and_finalize(
-            destination.as_path(),
-            FinalizeMetadataParams {
-                metadata: &metadata,
-                metadata_options,
-                mode,
-                path_context: MetadataPathContext {
-                    source: path_context.source.as_path(),
-                    relative: path_context.relative.as_deref(),
-                    file_type: path_context.file_type,
-                    destination_previously_existed: path_context.destination_previously_existed,
-                },
-                // upstream: receiver.c:964 - deferred updates have already
-                // committed the rename + applied dest_mode at the original
-                // commit site, so there is no pre-transfer stat to recover
-                // here.
-                pre_transfer_meta: None,
-                #[cfg(unix)]
-                fd: None, // No fd available for deferred updates
-                #[cfg(all(unix, feature = "xattr"))]
-                preserve_xattrs,
-                #[cfg(all(any(unix, windows), feature = "acl"))]
-                preserve_acls,
-            },
-        )
-    }
 
     /// Returns the configured delete timing (before, during, after, or delay).
     pub(super) const fn delete_timing(&self) -> Option<DeleteTiming> {
@@ -1516,78 +550,5 @@ impl<'a> CopyContext<'a> {
             info_log!(Nonreg, 1, "IO error encountered -- skipping file deletion");
         }
         true
-    }
-}
-
-/// Returns `true` when a hard-link or rename attempt into the backup area
-/// failed because a stale entry already occupies `backup_path`.
-///
-/// upstream: `backup.c:247` - `link_or_rename()` fails with `EEXIST` or
-/// `EISDIR` when the backup path is already occupied. Windows reports
-/// renaming or linking onto an existing directory as `ERROR_ACCESS_DENIED`
-/// (`PermissionDenied`) rather than `EEXIST`/`EISDIR`, so it is treated the
-/// same way; [`remove_stale_backup_entry`]'s re-stat still gates removal on
-/// the target actually existing, so a genuine permission error falls through
-/// to the retry-and-fail path unchanged.
-fn is_stale_backup_conflict(error: &io::Error) -> bool {
-    matches!(
-        error.kind(),
-        io::ErrorKind::AlreadyExists | io::ErrorKind::IsADirectory | io::ErrorKind::PermissionDenied
-    )
-}
-
-/// Clears whatever occupies `backup_path` so a hard-link or rename retry can
-/// land cleanly.
-///
-/// upstream: `backup.c:247-256` - `make_backup()` lstats the stale backup
-/// target and calls `delete_item(...DEL_FOR_BACKUP | DEL_RECURSE)` before
-/// retrying `link_or_rename()`.
-fn remove_stale_backup_entry(backup_path: &Path) -> Result<(), LocalCopyError> {
-    match fs::symlink_metadata(backup_path) {
-        Ok(meta) if meta.is_dir() => fs::remove_dir_all(backup_path).map_err(|remove_error| {
-            LocalCopyError::io("remove existing backup directory", backup_path, remove_error)
-        }),
-        Ok(_) => {
-            if let Err(remove_error) = fs::remove_file(backup_path)
-                && remove_error.kind() != io::ErrorKind::NotFound
-            {
-                return Err(LocalCopyError::io(
-                    "remove existing backup",
-                    backup_path,
-                    remove_error,
-                ));
-            }
-            Ok(())
-        }
-        Err(meta_error) if meta_error.kind() == io::ErrorKind::NotFound => Ok(()),
-        Err(meta_error) => Err(LocalCopyError::io(
-            "stat existing backup",
-            backup_path,
-            meta_error,
-        )),
-    }
-}
-
-/// Returns `true` when two nodes' modification times are equal within
-/// `--modify-window`.
-///
-/// upstream: util1.c:1478 same_time() - a whole-second delta within
-/// `modify_window` counts as unchanged. For a zero window (the default) or a
-/// negative window (`--modify-window < 0`, nsec-exact) a hardlink candidate
-/// must match on both the whole-second and nanosecond components.
-#[cfg(unix)]
-fn mtimes_within_window(
-    source: &fs::Metadata,
-    candidate: &fs::Metadata,
-    modify_window: ModifyWindow,
-) -> bool {
-    use std::os::unix::fs::MetadataExt;
-
-    let delta = source.mtime().abs_diff(candidate.mtime());
-    let window = modify_window.as_secs();
-    if window <= 0 {
-        delta == 0 && source.mtime_nsec() == candidate.mtime_nsec()
-    } else {
-        delta <= window as u64
     }
 }
